refactor(hook): 비동기 상태 커스텀 훅을 전체 공용 hook으로 이동
